+++ front/gatsby/src/components/BondingModal.tsx
@@ -2,11 +2,13 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
+import { DeriveFees } from '@polkadot/api-derive/types';
 import { SubmittableExtrinsic } from '@polkadot/api/types';
 import { createType } from '@polkadot/types';
+import { AccountInfo } from '@polkadot/types/interfaces';
 import {
   AccountsContext,
-  ApiContext,
+  ApiRxContext,
   ExtrinsicDetails,
   handler,
   TxQueueContext,
@@ -27,6 +29,7 @@ import {
 import BN from 'bn.js';
 import { navigate } from 'gatsby';
 import React, { useContext, useEffect, useState } from 'react';
+import { take } from 'rxjs/operators';
 
 import { validateFees } from '../util/validateExtrinsic';
 import { Button } from './Button';
@@ -64,37 +67,57 @@ const BondingModal = (): React.ReactElement => {
     currentAccount,
     loadingBalances,
   } = useContext(AccountsContext);
-  const { api, apiPromise, isApiReady } = useContext(ApiContext);
+  const { api, isApiReady } = useContext(ApiRxContext);
   const { enqueue, signAndSubmit, txQueue } = useContext(TxQueueContext);
 
   const [accountForController, setAccountForController] = useState(
     currentAccount
   );
   const [accountForStash, setAccountForStash] = useState(currentAccount);
+  const [accountNonce, setAccountNonce] = useState<AccountInfo>();
   const [allFees, setAllFees] = useState<BN>();
   const [allTotal, setAllTotal] = useState<BN>();
   const [bondAmount, setBondAmount] = useState<string>('');
   const [bondingError, setBondingError] = useState<Error>();
   const [extrinsic, setExtrinsic] = useState<SubmittableExtrinsic<'rxjs'>>();
+  const [fees, setFees] = useState<DeriveFees>();
   const [rewardDestination, setRewardDestination] = useState<RewardDestination>(
     RewardDestination.Staked
   );
   const [txId, setTxId] = useState<number>();
 
-  const checkFees = async () => {
+  useEffect(() => {
+    if (api && isApiReady) {
+      const sub = api.derive.balances
+        .fees()
+        .pipe(take(1))
+        .subscribe((result: DeriveFees) => setFees(result));
+
+      return () => sub.unsubscribe();
+    }
+  }, [api, isApiReady]);
+
+  useEffect(() => {
+    if (api && isApiReady) {
+      const sub = api.query.system
+        .account<AccountInfo>(accountForStash)
+        .pipe(take(1))
+        .subscribe((result: AccountInfo) => setAccountNonce(result));
+
+      return () => sub.unsubscribe();
+    }
+  }, [accountForStash, api, isApiReady]);
+
+  const checkFees = () => {
     if (
-      apiPromise &&
       isApiReady &&
       accountForStash &&
       accountForController &&
+      accountNonce &&
       bondAmount &&
-      extrinsic
+      extrinsic &&
+      fees
     ) {
-      const fees = await apiPromise.derive.balances.fees();
-      const accountNonce = await apiPromise.query.system.account(
-        accountForStash
-      );
-
       const [feeErrors, total, fee] = validateFees(
         accountNonce,
         new BN(bondAmount),
@@ -158,7 +181,7 @@ const BondingModal = (): React.ReactElement => {
   }, [currentAccount]);
 
   useEffect(() => {
-    if (apiPromise) {
+    if (api) {
       checkUserInputs();
       checkFees();
     }
@@ -166,7 +189,7 @@ const BondingModal = (): React.ReactElement => {
     accountForStash,
     accountForController,
     bondAmount,
-    apiPromise,
+    api,
     rewardDestination,
   ]);
 
@@ -194,7 +217,7 @@ const BondingModal = (): React.ReactElement => {
 
   const signAndSubmitBond = () => {
     if (
-      apiPromise &&
+      api &&
       accountForController &&
       accountForStash &&
       allFees &&