+++ front/gatsby/src/components/Layout/Header/Subheaders/EraHeader.tsx
@@ -3,7 +3,7 @@
 // of the Apache-2.0 license. See the LICENSE file for details.
 
 import { useQuery, useSubscription } from '@apollo/react-hooks';
-import { ApiContext } from '@substrate/context';
+import { ApiRxContext } from '@substrate/context';
 import React, { useContext, useEffect, useState } from 'react';
 
 import { ERAS_SUBSCRIPTION, LATEST_ERA_QUERY } from '../../../../util/graphql';
@@ -11,7 +11,7 @@ import HeaderItem from '../HeaderItem';
 import { EraHead } from '../types';
 
 const EraHeader = (): React.ReactElement => {
-  const { api } = useContext(ApiContext);
+  const { api } = useContext(ApiRxContext);
   const { data } = useSubscription(ERAS_SUBSCRIPTION);
   const queryData = useQuery(LATEST_ERA_QUERY);
   const [eraHead, setEraHead] = useState<EraHead>();