+++ front/gatsby/src/components/Layout/Header/Subheaders/StakingHeader.tsx
@@ -4,7 +4,7 @@
 
 import { useSubscription } from '@apollo/react-hooks';
 import { formatBalance } from '@polkadot/util';
-import { ApiContext } from '@substrate/context';
+import { ApiRxContext } from '@substrate/context';
 import React, { useContext, useEffect, useState } from 'react';
 
 import { STAKING_SUBSCRIPTION } from '../../../../util/graphql';
@@ -14,7 +14,7 @@ import { StakingHead } from '../types';
 const StakingHeader = (): React.ReactElement => {
   const { data } = useSubscription(STAKING_SUBSCRIPTION);
   const [stakeHead, setStakeHead] = useState<StakingHead>();
-  const { api } = useContext(ApiContext);
+  const { api } = useContext(ApiRxContext);
 
   useEffect(() => {
     if (data) {